# swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  Duration:
    description: |-
      A Duration represents the elapsed time between two instants
      as an int64 nanosecond count. The representation limits the
      largest representable duration to approximately 290 years.
    format: int64
    type: integer
    x-go-package: time
  ExpiringTask:
    properties:
      Complete:
        type: boolean
      CreatedAt:
        format: date-time
        type: string
      DeletedAt:
        format: date-time
        type: string
      Description:
        type: string
      ExpiringAt:
        format: date-time
        type: string
      ID:
        format: uint64
        type: integer
      PersonID:
        format: int64
        type: integer
      ProjectID:
        format: int64
        type: integer
      TimeLeft:
        $ref: '#/definitions/Duration'
      Title:
        type: string
      UpdatedAt:
        format: date-time
        type: string
    type: object
    x-go-package: github.com/psebaraj/gogetitdone/models
...
  Task:
    properties:
      Complete:
        type: boolean
      CreatedAt:
        format: date-time
        type: string
      DeletedAt:
        format: date-time
        type: string
      Description:
        type: string
      ID:
        format: uint64
        type: integer
      PersonID:
        format: int64
        type: integer
      ProjectID:
        format: int64
        type: integer
      Title:
        type: string
      UpdatedAt:
        format: date-time
        type: string
    type: object
    x-go-package: github.com/psebaraj/gogetitdone/models
info:
  description: Documentation of the GoGetItDone endpoints of the backend
  title: GoGetItDone
  version: 1.0.0
paths:
  /changedescrition/expiringtask/{id}:
    patch:
      consumes:
      - application/json
      operationId: changeExpiringTaskDescription
      produces:
      - application/json
      responses:
        "200":
          description: ExpiringTask
          schema:
            $ref: '#/definitions/ExpiringTask'
        "400":
          $ref: '#/responses/nil'
        "404":
          $ref: '#/responses/nil'
      tags:
      - ExpiringTask
...
  /changedescrition/task/{id}:
    patch:
      consumes:
      - application/json
      operationId: changeTaskDescription
      produces:
      - application/json
      responses:
        "200":
          description: Task
          schema:
            $ref: '#/definitions/Task'
        "400":
          $ref: '#/responses/nil'
        "404":
          $ref: '#/responses/nil'
      tags:
      - Task
  /changeproject/expiringtask/{id}:
    patch:
      consumes:
      - application/json
      operationId: changeExpiringTaskProject
      produces:
      - application/json
      responses:
        "200":
          description: ExpiringTask
          schema:
            $ref: '#/definitions/ExpiringTask'
        "400":
          $ref: '#/responses/nil'
        "404":
          $ref: '#/responses/nil'
      tags:
      - ExpiringTask
...
  /changeproject/task/{id}:
    patch:
      consumes:
      - application/json
      operationId: changeTaskProject
      produces:
      - application/json
      responses:
        "200":
          description: Task
          schema:
            $ref: '#/definitions/Task'
        "400":
          $ref: '#/responses/nil'
        "404":
          $ref: '#/responses/nil'
      tags:
      - Task
  /changetitle/expiringtask/{id}:
    patch:
      consumes:
      - application/json
      operationId: changeExpiringTaskTitle
      produces:
      - application/json
      responses:
        "200":
          description: ExpiringTask
          schema:
            $ref: '#/definitions/ExpiringTask'
        "400":
          $ref: '#/responses/nil'
        "404":
          $ref: '#/responses/nil'
      tags:
      - ExpiringTask
...
  /changetitle/task/{id}:
    patch:
      consumes:
      - application/json
      operationId: changeTaskTitle
      produces:
      - application/json
      responses:
        "200":
          description: Task
          schema:
            $ref: '#/definitions/Task'
        "400":
          $ref: '#/responses/nil'
        "404":
          $ref: '#/responses/nil'
      tags:
      - Task
  /complete/expiringtask/{id}:
    patch:
      operationId: toggleExpiringTaskComplete
      produces:
      - application/json
      responses:
        "200":
          description: ExpiringTask
          schema:
            $ref: '#/definitions/ExpiringTask'
        "400":
          $ref: '#/responses/nil'
        "404":
          $ref: '#/responses/nil'
      tags:
      - ExpiringTask
...
  /complete/task/{id}:
    patch:
      operationId: toggleTaskComplete
      produces:
      - application/json
      responses:
        "200":
          description: Task
          schema:
            $ref: '#/definitions/Task'
        "400":
          $ref: '#/responses/nil'
        "404":
          $ref: '#/responses/nil'
      tags:
      - Task
  /create/expiringtask:
    post:
      consumes:
      - application/json
      operationId: createExpiringTask
      produces:
      - application/json
      responses:
        "201":
          description: ExpiringTask
          schema:
            $ref: '#/definitions/ExpiringTask'
        "507":
          $ref: '#/responses/nil'
      tags:
      - ExpiringTask
...
  /delete/expiringtask/{id}:
    delete:
      operationId: deleteExpiringTask
      produces:
      - application/json
      responses:
        "200":
          description: ExpiringTask
          schema:
            $ref: '#/definitions/ExpiringTask'
        "404":
          $ref: '#/responses/nil'
        "500":
          $ref: '#/responses/nil'
      tags:
      - ExpiringTask
...
  /expiringtask/{id}:
    get:
      operationId: getExpiringTask
      produces:
      - application/json
      responses:
        "200":
          description: ExpiringTask
          schema:
            $ref: '#/definitions/ExpiringTask'
        "404":
          $ref: '#/responses/nil'
      tags:
      - ExpiringTask
  /expiringtasks:
    get:
      operationId: getExpiringTasks
      produces:
      - application/json
      responses:
        "200":
          description: ExpiringTask
          schema:
            items:
              $ref: '#/definitions/ExpiringTask'
            type: array
        "204":
          $ref: '#/responses/nil'
      tags:
      - ExpiringTask
...
produces:
- application/json
...
schemes:
- http
swagger: "2.0"